[!U] Enabled climbing up a wall function.
Climbing up a wall function has been completed. Currently the player can climb a wall that is in the direction of movement. This is the base functionality of the character. Next step is to implement the ability to turn 90 degrees towards the back wall to climb. This will give the level designer greater functionality for the levels.