[!U] Climbing background objects is now enabled
Climbing background objects is now enabled. When the character is facing on the Y the character successfully disconnects from the wall and falls down. When the character is facing the negative Y, as it approaches the top of the wall it rotates to the positive Y and gets stuck. Check the Dot Product of the climbing function.