[!U] CustomMovementComponent function was changed to allow the character to detach from the top of the ledge when facing either direction.
The character now detaches from the ledge and does not get stuck. Character collision component was getting stuck in the World Static object. The character was moving towards the collision object when it was facing along the Negative Y Direction. This had to be negated for the character to not snap to the wall.