[!U] Cleaned up and organized the AnimBP with the proper parent class, variables and added a Fall Loop animation.
All the animations that have been added so far are just placeholder animations and will be replaced at a later time. The next step is to setup the enter and exit state for the bottom of the ladder animations. The functionality to enable and disable the ladder with the character is implemented. The animations are not enabled yet.