[!U] Foot and Hand IK offsets when climbing is enabled.
The Character Rig has offsets for the hands and feet. These are setup in the Control Rig and will have to be input manually. This will ensure the animations do not need to be adjusted to the ladder itself, the offset will take care of the hand and feet placement.
The MotionWarping plugin has been enabled for the Vaulting movement.