[!A] Setup Blend space for Unarmed Movement
Blend Space created and added to animation BP. Currently using BP_Manny as the Animation Blueprint with the Quinn mesh. This will changed when we have a custom character BO in the future. The Blend Space replaced the default Walk/Run animation in the ABP_Manny and an additional setup was created in the Event Graph. See comments in ABP_Manny. Nine Unarmed animations were added to the project.